% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootCImic.R
\name{bootCImic}
\alias{bootCImic}
\title{Bootstrap confidence intervals for MIC}
\usage{
bootCImic(
  data,
  b = 1000,
  mic = c("roc", "predict", "adjust"),
  model = NULL,
  ...
)
}
\arguments{
\item{data}{data.frame to bootstrap}

\item{b}{number of bootstrap samples}

\item{mic}{mic method to apply bootstraps for \code{c("roc", "predict", "adjust")}}

\item{model}{only used for \code{mic = "adjust"} the model described using the
\code{lavaan} model syntax. see examples in \link{tr_reliability} or \code{?lavaan::model.syntax}}

\item{x}{vector with score at time point 1, or column name in the data if
\code{!is.null(data)}}

\item{y}{vector with score at time point 2, or column name in the data if
\code{!is.null(data)}}

\item{tr}{vector with transition rates (perceived change), or column name in
the data if \code{!is.null(data)}}
}
\value{
a tibble with columns .lower, .estimate, .upper, .alpha, .method, and
term. The term is the mic method.
}
\description{
Bootstrap confidence intervals for MIC
}
\examples{
nitems <- 10
example$x <- rowSums(example[,1:nitems])                 # sumscore T1
example$y <- rowSums(example[,(nitems+1):(2*nitems)])    # sumscore T2
model <- '
# Factors
F1 =~ a1*v1_1+a2*v1_2+a3*v1_3+a4*v1_4+a5*v1_5+a6*v1_6+a7*v1_7+a8*v1_8+
  a9*v1_9+a10*v1_10+trat
F2 =~ a1*v2_1+a2*v2_2+a3*v2_3+a4*v2_4+a5*v2_5+a6*v2_6+a7*v2_7+a8*v2_8+
  a9*v2_9+a10*v2_10+trat
  '
#start <- Sys.time()
#bootCImic(example, model = model)
#Sys.time()-start

start <- Sys.time()
bootCImic(example, mic = c("roc", "pred"))
Sys.time()-start

}
