% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIC_helpers.R
\name{clust_align}
\alias{clust_align}
\title{Cluster Aligner}
\usage{
clust_align(Z1, Z2, type = "vec")
}
\arguments{
\item{Z1}{Clustering assignment No.1 (reference).}

\item{Z2}{Clustering assignment No.2 (to be adjusted).}

\item{type}{Option of '\emph{vec}' and '\emph{mat}' as input format.}
}
\value{
\code{Z2} aligned against \code{Z1} achieving maximal concordance.
}
\description{
\code{clust_align} aligns two cluster assignments by relabelling the membership for
  a maximized concordance
}
\details{
This aligner can take both \emph{vector} and \emph{matrix} clustering assignments, and aligns
  the second assignment maximally against the first assignment. It aims at reshuffling of
  cluster labels to facilitate a fair comparision between two clustering result.
}
\examples{
Z1 <- c(1:5)
Z2 <- c(5:1)

clust_align(Z1, Z2, type = 'vec')

}

