% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patric.R
\name{download_patric_db}
\alias{download_patric_db}
\title{Download PATRIC database}
\usage{
download_patric_db(save_path, ftp_path = patric_ftp_path, overwrite = FALSE)
}
\arguments{
\item{save_path}{Save path (should be .txt)}

\item{ftp_path}{PATRIC database FTP path to download}

\item{overwrite}{Force overwrite}
}
\value{
TRUE if successful, FALSE if failure.
}
\description{
Download PATRIC database
}
\examples{
\donttest{
download_patric_db(tempfile())
}
}
