% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoffs}
\alias{ecoffs}
\title{ECOFF data}
\format{
\subsection{\code{ecoffs}}{

A data frame with 85 rows and 25 columns:
\describe{
\item{organism}{Microorganism code in AMR::mo format}
\item{antibiotic}{Antibiotic code in AMR::ab format}
\item{\code{0.002}:\code{512}}{Counts of isolates in each concentration "bin"}
\item{Distributions}{see EUCAST documentation below}
\item{Observations}{Number of observations}
\item{\verb{(T)ECOFF}}{see EUCAST documentation below}
\item{\verb{Confidence interval}}{see EUCAST documentation below}
}
}
}
\source{
EUCAST \url{https://www.eucast.org/mic_and_zone_distributions_and_ecoffs}

These data have (or this document, presentation or video has) been produced
in part under ECDC service contracts and made available by EUCAST at no cost
to the user and can be accessed on the EUCAST website www.eucast.org.
The views and opinions expressed are those of EUCAST at a given point in time.
EUCAST recommendations are frequently updated and the latest versions are available at www.eucast.org.
}
\usage{
ecoffs
}
\description{
A dataset containing the epidemiological cut-off values (ECOFFs) for
different antibiotics and microorganisms. Currently, only the ECOFF values
for \emph{Escherichia coli} are included.
}
\keyword{datasets}
