% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_lap.R
\name{eigen_lap}
\alias{eigen_lap}
\title{Eigen-Laplacian Dimension Reduction}
\usage{
eigen_lap(X, D)
}
\arguments{
\item{X}{Output \code{diss_array} from \code{\link{ep_dismat}}, featuring pairwise distances on epoch-smoothed units.}

\item{D}{Dimensionality of the output, which correspond to the number of eigenvectors extracted from graph Laplacian.}
}
\value{
A list of objects with the following components:
  \item{\code{eig_data}}{A 3-D array of the Eigen-Laplacian representation.}
  \item{\code{eig_value}}{Corresponding eigen-values of the extracted eigen-vectors}
}
\description{
\code{eigen_lap} carries out eigen-decomposition on the Laplacian matrices, which
  operates on distance matrices and yields low dimensional representations of original objects.
}
\details{
This procedure is a direct implementation of the spectral clustering technique explicated by [Ng 2001].
It is also a preprocessing step for \code{\link{MIC}}, following distance characterization by \code{\link{ep_dismat}}.
}
\examples{
Later on

}
\references{
Ng, Andrew Y., Michael I. Jordan, and Yair Weiss. \emph{On spectral clustering: Analysis and an algorithm.}
  Advances in neural information processing systems 2 (2002): 849-856.
}

