% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_dismat.R
\name{ep_dismat}
\alias{ep_dismat}
\title{Epoch-wise pairwise Distance Matrices}
\usage{
ep_dismat(X, sf, par.spectrum = c(100, sf/2, 512), window = c(1, 0))
}
\arguments{
\item{X}{A 3-D array of input organizing the data in the stucture of:
No.Objects * No.Observations * No.Segments. Here objects refer to the clustering units;
observations refer to the time series at each segment; and segments are pieces of time series
that are repeatedly observed/measured of same length.}

\item{sf}{Sampling frequency}

\item{par.spectrum}{Parameter for \code{\link{spec.parzen}} in the order of \code{a, wn, nn}.}

\item{window}{Epoch-smoothing window configuration, in the order of (window size, overlap size).
For example, non-smoothing setting makes \code{window = c(1,0)}.}
}
\value{
a list object with the following components:
  \item{\code{diss_array}}{Pairwise distance array, No.objects*No.objects*No.epochs}
  \item{\code{ave_spec}}{Epochwise average spectral estimates using \code{\link{spec.parzen}}}
  \item{\code{fw}}{Frequencies used for spectral estimate}
}
\description{
\code{ep_dismat} calculates strucally smoothed pairwise distances across all objects.
  It is an essential prerequisite step of MIC implementation.
}
\details{
This procedure consists of distance characterization across all objects to be clustered on
  and further smoothing into epochs.

The distance metric is based on the \code{\link{TVD}} between a pair of spectral
  densities, calculated by the Parzen lag window estimator \code{\link{spec.parzen}}.

Epochs are constructed by combining adjacent units (segments), which are further allowed
  to overlap with each other at certain amount.
}
\examples{
Later on

}

