% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_mics}
\alias{example_mics}
\title{Example MIC data}
\format{
\subsection{\code{example_mics}}{

A data frame with 300 rows and 4 columns:
\describe{
\item{gs}{Gold standard MICs}
\item{test}{Test MICs}
\item{mo}{Microorganism code in AMR::mo format}
\item{ab}{Antibiotic code in AMR::ab format}
}
}
}
\source{
Synthetic data
}
\usage{
example_mics
}
\description{
Example minimum inhibitory concentration validation data for three
antimicrobials on Escherichia coli strains. This data is synthetic and
generated to give an example of different MIC distribution.
}
\keyword{datasets}
