% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIC_helpers.R
\name{HardCluster}
\alias{HardCluster}
\title{Hard Clustering from MCMC samples}
\usage{
HardCluster(ClustList)
}
\arguments{
\item{ClustList}{matrix, with row vectors being cluster labels that are properly aligned.
Usually organized as No.iterations * No.objects.}
}
\value{
A list of objects with the following components:
  \item{\code{Cbest}}{vector, hard clustering estimate}
  \item{\code{vec}}{vector, Frobenius norm to the mean affinity (analog of deviation from the mean)}
}
\description{
\code{HardCluster} produces a point clustering estimate from a sample clustering labels (MCMC outputs)
}
\details{
This procedure treats each clustering assignments as affinity matrices, and seek for a point estimate
  that minimizes the Frobenius norm to mean affinity matrix, from the assignments being traversed by the
  sample.
}
\seealso{
\code{\link{MIC}} for its usage in MCMC, \code{\link{clust_align}} for clusters alignment.
}

