% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIC_helpers.R
\name{logSum}
\alias{logSum}
\title{MIC Helper: logSum}
\usage{
logSum(l)
}
\arguments{
\item{l}{vector}
}
\value{
numeric
}
\description{
\code{logSum} is a samll helper function calculating the sum of logarithms.
}
\seealso{
\code{\link{MIC}} for its usage
}

