% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIC_helpers.R
\name{margPi}
\alias{margPi}
\title{MIC Helper: margPi}
\usage{
margPi(p, a)
}
\arguments{
\item{p}{vector of proportions, sum up to 1}

\item{a}{numeric, adherence}
}
\value{
numeric
}
\description{
\code{margPi} is a samll helper function calculating the marginal prior of MIC.
}
\seealso{
\code{\link{MIC}} for its usage
}

