% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIC_prep.R
\name{MIC_prep}
\alias{MIC_prep}
\title{Data preparation for MIC}
\usage{
MIC_prep(X, d, exclude = c(), par.spectrum = c(50), par.win = c(1, 0))
}
\arguments{
\item{X}{3D array, organized as No.objects * No.observations * No.segments.}

\item{d}{Dimensionality of the eigen-Laplacian representation.}

\item{exclude}{Vector of integers, objects to be excluded.}

\item{par.spectrum}{Vector, parameters for \code{\link{spec.parzen}}}

\item{par.win}{Vector, epoch smoothing parameters for \code{\link{ep_dismat}}}
}
\value{
List of data matrices, each with No.objects rows and \code{d} columns.
}
\description{
\code{MIC_prep} carries out signal detrending, epoch smoothing, distance characterization and dimensional reduction sequentiall, therefore
 prepares the structured data for MIC analysis.
}
\examples{
\dontrun{
# Simulated data:
ts_sim <- MIC_sim(alpha = 0.9, nsub = 3, segs = 10, fs = 100)$Data

# Data preparation, subject 1 epoch 1

sub1 <- MIC_prep(ts_sim[[1]], d = 3, par.spectrum = c(50, 50), par.win = c(3, 1))

# No. of epochs

length(sub1)

# Epoch level data

dim(sub1[[1]])
}
}
\seealso{
\code{\link{spec.parzen}} for spectral estimate, \code{\link{ep_dismat}} for epoch smoothing, \code{\link{eigen_lap}} for eigen-Laplacian
 and \code{\link{MIC_sim}} for time series simulation.
}

