% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mic_manipulation.R
\name{mic_range}
\alias{mic_range}
\title{Generate dilution series}
\usage{
mic_range(start = 512, dilutions = Inf, min = 0.002, precise = FALSE)
}
\arguments{
\item{start}{starting (highest) concentration}

\item{dilutions}{number of dilutions}

\item{min}{minimum (lowest) concentration}

\item{precise}{force range to be high precision (not usually desired
behaviour)}
}
\value{
Vector of numeric concentrations
}
\description{
Generate dilution series
}
\examples{
mic_range(128)
mic_range(128, dilutions = 21) # same results
}
