% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mim_threshold.R
\name{mim_threshold_details}
\alias{mim_threshold_details}
\title{Extract hidden details from a mim_threshold object}
\usage{
mim_threshold_details(x)
}
\arguments{
\item{x}{A \code{mim_threshold} object.}
}
\value{
A list containing hidden details.
}
\description{
Extract hidden details from a mim_threshold object
}
