% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{print.mic_validation}
\alias{print.mic_validation}
\title{Print MIC validation object}
\usage{
\method{print}{mic_validation}(x, ...)
}
\arguments{
\item{x}{mic_validation object}

\item{...}{additional arguments}
}
\value{
character
}
\description{
Print MIC validation object
}
\examples{
gold_standard <- c("<0.25", "8", "64", ">64")
test <- c("<0.25", "2", "16", "64")
val <- compare_mic(gold_standard, test)
print(val)
}
