% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_converters.R
\name{replace_multiple_slashes}
\alias{replace_multiple_slashes}
\title{Removes multiple slashes in a path or url}
\usage{
replace_multiple_slashes(path)
}
\arguments{
\item{path}{character vector}
}
\value{
character vector of paths without duplicate slashes
}
\description{
Removes multiple slashes in a path or url
}
