% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/kmer_batch.R
\name{reverse_complement}
\alias{reverse_complement}
\title{Reverse complement of DNA string}
\usage{
reverse_complement(dna)

reverse_complement(dna)
}
\arguments{
\item{dna}{DNA string}
}
\value{
reverse complement of DNA string
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
reverse_complement("ATCG")
}
