% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_threshold.R
\name{sim_threshold_details}
\alias{sim_threshold_details}
\title{Extract details from a SIM threshold object}
\usage{
sim_threshold_details(x)
}
\arguments{
\item{x}{A \code{sim_threshold} object.}
}
\value{
A list of additional model details.
}
\description{
\code{sim_threshold_details()} extracts additional details from an object returned
by \code{sim_threshold()}, including the fitted lavaan model, CFA data,
latent anchor location, model-implied category probabilities, and bootstrap
results.
}
