% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat.R
\name{simdat}
\alias{simdat}
\title{Simulates Time 1 and Time 2 PROM items and a transition rating vector}
\usage{
simdat(
  N = 2000,
  par.mn.imic = 0.37425,
  par.sd.imic = 0.05,
  rt1ch = -0.5,
  mean.tetchs = 0.3,
  par.sd.tetchs = 1,
  par.rel.trt = 0.7,
  seed = 1234
)
}
\arguments{
\item{N}{Integer. Sample size for simulation.}

\item{par.mn.imic}{Numeric. Mean individual MIC. When 0.37425, corresponds
approximately to MIC 2.5 points on the raw scale.}

\item{par.sd.imic}{Numeric. Standard deviation of individual MICs.}

\item{rt1ch}{Numeric. Correlation between baseline theta and latent change.}

\item{mean.tetchs}{Numeric. Mean latent change.}

\item{par.sd.tetchs}{Numeric. Standard deviation of latent change.}

\item{par.rel.trt}{Numeric. Transition rating reliability.}

\item{seed}{Optional integer. Random seed used to make the simulated data
reproducible. If \code{NULL}, the current random-number generator state is used.}
}
\value{
A list of simulation summaries and a data frame with Time 1 items,
Time 2 items, transition rating \code{trat}, and observed PROM change score
\code{xoc}.
}
\description{
Simulates Time 1 and Time 2 PROM item responses and a transition rating
vector from an item response theory context.
}
\details{
R code adapted from supplementary materials of Terluin et al.
Qual Life Res. 2024;33:963–973.
}
