% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/kmer_batch.R
\name{squeezed_mers}
\alias{squeezed_mers}
\title{Generates all permutations of squeezed kmers}
\usage{
squeezed_mers(k = 3L)

squeezed_mers(k = 3L)
}
\arguments{
\item{k}{kmer length}
}
\value{
vector of squeezed kmers
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
squeezed_mers(3)
}
