% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{TVD}
\alias{TVD}
\title{Total Variation Distance}
\usage{
TVD(w, f1, f2)
}
\arguments{
\item{w}{A vector of sampled points.}

\item{f1}{Density 1 evaluated at values \code{w}.}

\item{f2}{Density 2 evaluated at values \code{w}.}
}
\value{
A scalar, in range [0,1] if both \code{f1} and \code{f2} are normalized densities.
}
\description{
\code{TVD} returns the Total Variation Distance between two densities.
}
\examples{
w <- seq(0, 1, by =.05)
f1 <- dbeta(w, 2, 2); f2 <- dbeta(w, 2, 5)

TVD(w, f1, f2)
TVD(w, f2, f1)
}

