% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/kmer_batch.R
\name{unsqueezed_index_to_str}
\alias{unsqueezed_index_to_str}
\title{Get str conversion of unsqueezed kmer using index}
\usage{
unsqueezed_index_to_str(x, k, starting_index = 1L)

unsqueezed_index_to_str(x, k, starting_index = 1L)
}
\arguments{
\item{x}{integer vector of kmer indices}

\item{k}{kmer length}

\item{starting_index}{starting index (libsvm is usually indexed starting at 1)}
}
\value{
vector of unsqueezed kmer strings
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
unsqueezed_index_to_str(2, k = 3)
}
