% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/kmer_batch.R
\name{unsqueezed_mers}
\alias{unsqueezed_mers}
\title{Generates all permutations of unsqueezed kmers}
\usage{
unsqueezed_mers(k = 3L)

unsqueezed_mers(k = 3L)
}
\arguments{
\item{k}{kmer length}
}
\value{
vector of unsqueezed kmers
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
unsqueezed_mers(3)
}
