% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIC_helpers.R
\name{clust_align}
\alias{clust_align}
\title{Cluster Aligner}
\usage{
clust_align(Z1, Z2, type = "vec")
}
\arguments{
\item{Z1}{vector or matrix, clustering assignment 1 (reference).}

\item{Z2}{vector or matrix, clustering assignment 2 (to be relabeled).}

\item{type}{what type of input format: "\code{vec}" or "\code{mat}"}
}
\value{
\code{Z2}, aligned against \code{Z1} with maximal concordance.
}
\description{
\code{clust_align} aligns two cluster assignments by a maximally concordant relabeling
}
\details{
This aligner can take both \emph{vector} and \emph{matrix} clustering assignments, and aligns
  the second assignment against the first assignment. It reshuffles the second assignment labels
  such that two assignemnts are maximally matched. This procedure facilitates a direct and
  fair comparision between two clustering result.
}
\examples{
Z1 <- c(1:5)
Z2 <- c(5:1)

clust_align(Z1, Z2, type = 'vec')

}

