% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\docType{data}
\name{example}
\alias{example}
\title{example}
\format{
A data frame with 1000 rows and 21 variables:
\describe{
\item{v1_1}{item1 at baseline}
\item{v1_2}{item2 at baseline}
\item{v1_3}{item3 at baseline}
\item{v1_4}{item4 at baseline}
\item{v1_5}{item5 at baseline}
\item{v1_6}{item6 at baseline}
\item{v1_7}{item7 at baseline}
\item{v1_8}{item8 at baseline}
\item{v1_9}{item9 at baseline}
\item{v1_10}{item10 at baseline}
\item{v2_1}{item1 at repeated measurement}
\item{v2_2}{item2 at repeated measurement}
\item{v2_3}{item3 at repeated measurement}
\item{v2_4}{item4 at repeated measurement}
\item{v2_5}{item5 at repeated measurement}
\item{v2_6}{item6 at repeated measurement}
\item{v2_7}{item7 at repeated measurement}
\item{v2_8}{item8 at repeated measurement}
\item{v2_9}{item9 at repeated measurement}
\item{v2_10}{item10 at repeated measurement}
\item{trat}{transition rating, dichotomous indicator}
}
}
\source{
data-raw/R/example.R
}
\usage{
example
}
\description{
Simulated example data of 10 items measured at two timepoints for 1000 subjects.
}
\keyword{datasets}
