% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIC_helpers.R
\name{HardCluster}
\alias{HardCluster}
\title{Hard Clustering from MCMC samples}
\usage{
HardCluster(ClustList)
}
\arguments{
\item{ClustList}{A matrix of clustering results whose row vectors are assignments, that have been properly aligned.
Usually organized as No.Iterations * No.Objects.}
}
\value{
A list of objects with the following components:
  \item{\code{Cbest}}{Hard Clustering point estimate}
  \item{\code{vec}}{A vector of Frobenius norm to the mean affinity}
}
\description{
\code{HardCluster} produces a point clustering estimate from a sample of posterior estimates (MCMC outputs)
}
\details{
This procedure treats each clustering assignments as affinity matrices, and seek for a point estimate
  that minimizes the Frobenius norm to mean affinity matrix, from the assignments being traversed by the
  sample.
}
\seealso{
\code{\link{MIC}} for its usage in MCMC, \code{\link{clust_align}} for clusters alignment.
}

