% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIC-package.R
\docType{package}
\name{MIC-package}
\alias{MIC}
\alias{MIC-package}
\title{MIC: Analysis of Antimicrobial Minimum Inhibitory Concentration Data}
\description{
Analyse, plot, and tabulate antimicrobial minimum inhibitory concentration (MIC) data. Validate the results of an MIC experiment by comparing observed MIC values to a gold standard assay, in line with standards from the International Organization for Standardization (2021) \url{https://www.iso.org/standard/79377.html}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/agerada/MIC}
  \item Report bugs at \url{https://github.com/agerada/MIC/issues}
}

}
\author{
\strong{Maintainer}: Alessandro Gerada \email{alessandro.gerada@liverpool.ac.uk} (\href{https://orcid.org/0000-0002-6743-4271}{ORCID}) [copyright holder]

}
\keyword{internal}
