% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIC.R
\name{MIC}
\alias{MIC}
\title{Multilevel Integrative Clustering}
\usage{
MIC(X, K, NumRun, a = 1, b = 1, IndivAlpha = TRUE, IndivBeta = TRUE,
  Concentration = 1, adj = 5e-05, block = 1000, time.start = Sys.time())
}
\arguments{
\item{X}{list of lists, and each sublist contains data matrices of the same size, for a proper format see
the output of \code{\link{MIC_prep}}}

\item{K}{integer, number of clusters}

\item{NumRun}{number of iterations for MCMC sampling, with 1/5 used for burn-in}

\item{a}{numeric, hyperparameter for adherence ~ TBeta(a, b)}

\item{b}{numeric, hyperparameter for adherence ~ TBdta(a, b)}

\item{IndivAlpha}{boolean, whether to assume individual adherence (population - subjects)}

\item{IndivBeta}{boolean, whether to assume individual adherence (subjects - epochs)}

\item{Concentration}{numeric, hyperparameter for Dirichlet prior}

\item{adj}{numeric, adjustment term for label sampling with default of \code{5e-05}}

\item{block}{integer, unit size of MCMC samples to be written externally (\strong{memory saving!})}

\item{time.start}{the start of code timing}
}
\value{
A list of objects with the following components:
  \item{\code{Alpha, AlphaBounds}}{point estimates and CIs of subject adherence to a group estimate}
  \item{\code{Betas, BetaBounds}}{point estimates and CIs of epoch adherence to its subject estimate}
  \item{\code{Cbest}}{point estimate of group level clustering result}
  \item{\code{Cibest}}{point estiamte of subject level clustering result}
  \item{\code{Lbest}}{point estimates of epoch level clustering result}
  \item{\code{ICs}}{model assessment metrics: BIC, BIC2, DIC4, DIC7, adjusted coherence}
  \item{\code{...}}{variability measures, posterior clustering probability matrices, etc. Please see our manuscript for details}
}
\description{
\code{MIC} implements integrative clustering on highly structured data. The admissible structure contains repeated measurements
  from multiple subjects, where both a group clustering and individual clusterings are of inferential interests.
}
\examples{

\dontrun{
  # Time Series simulation:
  ts_sim <- MIC_sim(alpha = 0.9, nsub = 3, segs = 10, fs = 100)


  # Data preparation:
  list_data <- list()
  for (i in 1:3) list_data[[i]] <- MIC_prep(X = ts_sim$Data[[i]], d = 4,
    par.spectrum = c(50, 50, 100), par.win = c(3, 1))



  # MIC: (Running time: 3mins)
  output <- MIC(X = list_data, K = 4, NumRun = 5000)



  # Clustering accuracy: group level
  sum(clust_align(ts_sim$C, output$Cbest, type = 'vec') == ts_sim$C) / 40
  ## [1] 1


  # Clustering accuracy: individual level
  for (i in 1:3){
  aligned_label <- clust_align(ts_sim$Ci[, i], output$Cibest[, i], type = 'vec')
  cat(sum(aligned_label == ts_sim$Ci[, i]) / 40, '\\n')
  }
  ## 1 1 1
}

}
\references{
Qian Li, Damla Senturk, Catherine A. Sugar, Shanali Jeste, Charlotte DiStefano, Joel Frohlich, Donatello Telesca
  "\emph{Inferring Brain Signals Synchronicity from a Sample of EEG Readings}".
}
\seealso{
\code{\link{eigen_lap}}, \code{\link{ep_dismat}} and \code{\link{MIC_prep}} for data preparation, \code{\link{spec.parzen}} for the estimation
  of spectrum and \code{\link{TVD}} for pairwise distance metric, \code{\link{HardCluster}} for point clustering estimate from
  MCMC samples.
}

