% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mic_iapm.R
\name{mic_iapm}
\alias{mic_iapm}
\title{Predictive modeling-based, adjusted predictive modeling-based, and "improved" adjusted predictive modeling-based MICs and their attendant bootstrapped 95\% CIs
This function could also be used to estimate the interpretation threshold of a predictor}
\usage{
mic_iapm(
  mypred,
  anchor,
  mydata,
  tr_rel,
  nboot = 0,
  report_every = 100,
  verbose = FALSE
)
}
\arguments{
\item{mypred}{Character vector; name of the column containing the change score or predictor score.}

\item{anchor}{Character vector; name of the column containing the transition rating (perceived change).}

\item{mydata}{Data.frame with the change score (or predictor score) and the transition rating in separate columns.}

\item{tr_rel}{Reliability of the transition rating estimated using \code{tr_reliability}.}

\item{nboot}{Integer; number of bootstrapped samples when estimating the 95\% CI.  \code{nboot} must  > 100 to trigger the bootstrapping process.}

\item{report_every}{Integer. The interval at which the counter should be printed.}

\item{verbose}{Logical. If TRUE, the counter is printed every \code{reoprt_every} iterations. If FALSE, no counter is printed.}
}
\value{
A list of Predictive modeling-based MIC estimates
}
\description{
Based on methods developed by Terluin et al (2015), Terluin et al (2017), and Terluin et al (2022).
}
\examples{


drop additional classes (e.g. tbl_df and data.table
progress indicator
prevalence adjustment
MICs
bootstrapped CIs for APM or iAPM
bootstrapping initiated when nboot > 100
}
\seealso{
Terluin B, Eekhout I, Terwee CB, de Vet HCW. Minimal important change (MIC) based on a predictive modeling approach was more precise than MIC based on ROC analysis. J Clin Epidemiol. 2015;68(12):1388-1396. doi:10.1016/j.jclinepi.2015.03.015

Terluin B, Eekhout I, Terwee CB. The anchor-based minimal important change, based on receiver operating characteristic analysis or predictive modeling, may need to be adjusted for the proportion of improved patients. J Clin Epidemiol. 2017;83:90-100. doi:10.1016/j.jclinepi.2016.12.015

Terluin B, Eekhout I, Terwee CB. Improved adjusted minimal important change took reliability of transition ratings into account. J Clin Epidemiol. 2022;148:48-53. doi:10.1016/j.jclinepi.2022.04.018
}
