% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mic_lcfa.R
\name{mic_lcfa}
\alias{mic_lcfa}
\title{Estimates (i) present state bias in transition ratings and (ii) anchor-based minimal important change using longitudinal confirmatory factor analysis}
\usage{
mic_lcfa(mydat, model, trt_cut = 1, B = 0, report_every = 100)
}
\arguments{
\item{mydat}{A data frame in which the variables are ordered to include Time 1 items, Time 2 items, and a single transition rating variable. User needs to ensure that each PROM item has the same number of levels at both time-points.}

\item{model}{Model specification based on \code{lavaan} model syntax}

\item{trt_cut}{Numeric vector; cutpoint to define a dichotomized transition rating variable}

\item{B}{Integer; number of bootstrapped samples. \code{B}>0 triggers the bootstrapping process.}

\item{report_every}{Integer. The interval at which the counter should be printed.}
}
\value{
a list of estimates including the present state bias and the MIC with its bootstrapped 95\% CI.
}
\description{
Based on methods developed by Terluin et al J Clin Epidemiol. 2022;143:128-136. and Qual Life Res. 2024;33:963–73
}
\examples{
single factor IRT model
single factor IRT model
}
\seealso{
Terluin B, Trigg A, Fromy P, Schuller W, Terwee CB, Bjorner JB. Estimating anchor-based minimal
important change using longitudinal confirmatory factor analysis. Qual Life Res. 2024;33:963–73.
Terluin B, Griffiths P, Trigg A, Terwee CB, Bjorner JB.
Present state bias in transition ratings was accurately estimated in simulated and real data. J Clin Epidemiol. 2022;143:128-136. doi:10.1016/j.jclinepi.2021.12.024
}
