% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mic_pred.R
\name{mic_pred}
\alias{mic_pred}
\title{Predicted Minimal Important Change (MIC)}
\usage{
mic_pred(data = NULL, x, y, tr)
}
\arguments{
\item{data}{data.frame that holds the scores at two time points and the
transition rate in separate columns.}

\item{x}{vector with score at time point 1, or column name in the data if
\code{!is.null(data)}}

\item{y}{vector with score at time point 1, or column name in the data if
\code{!is.null(data)}}

\item{tr}{vector with transition rates (perceived change), or column name in
the data if \code{!is.null(data)}}
}
\value{

}
\description{
Minimal Important Change (MIC) obtained as predicted value from a logistic
regression model.
}
\examples{
nitems <- 10
example$x <- rowSums(example[,1:nitems])                 # sumscore T1
example$y <- rowSums(example[,(nitems+1):(2*nitems)])    # sumscore T2
mic_pred(x = example$x, y = example$y, tr = example$trat)
mic_pred(data = example, x = example$x, y = example$y, tr = example$trat)
mic_pred(data = example, x = "x", y = "y", tr = "trat")
mic_pred(data = example, x = "x", y = "z", tr = "trat")
}
