% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reverse_complement}
\alias{reverse_complement}
\title{Reverse complement of DNA string}
\usage{
reverse_complement(dna)
}
\arguments{
\item{dna}{DNA string}
}
\value{
reverse complement of DNA string
}
\description{
Reverse complement of DNA string
}
\examples{
reverse_complement("ATCG")
}
