% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat.R
\name{simdat}
\alias{simdat}
\title{Simulates Time 1 and Time 2 PROM items and a transition rating vector from an item response theory context.}
\usage{
simdat(
  N = 2000,
  par.mn.imic = 0.37425,
  par.sd.imic = 0.05,
  rt1ch = -0.5,
  mean.tetchs = 0.3,
  par.sd.tetchs = 1,
  par.rel.trt = 0.7
)
}
\arguments{
\item{N}{sample size for simulation}

\item{par.mn.imic}{Numeric vector; mean individual MICs. When 0.37425, corresponds to MIC 2.5 points on the raw scale.}

\item{par.sd.imic}{Numeric vector; sd of individual MICs. Cannot be too large lest MIC is negative}

\item{rt1ch}{Numeric vector; correlation between tet1s and latent change}

\item{mean.tetchs}{Numeric vector; mean latent change}

\item{par.sd.tetchs}{Numeric vector; standard deviation of \code{tetchs}}

\item{par.rel.trt}{Numeric vector; transition rating reliability}
}
\value{
a list of estimates and a dataframe with Time 1 and Time 2 PROM items and a transition rating vector (\code{trat})
}
\description{
R codes taken from supplementary materials of Terluin et al Qual Life Res. 2024;33:963–73
}
\examples{
Simulate baseline data-set and change (theta) scores
generate output
}
\seealso{
Terluin B, Trigg A, Fromy P, Schuller W, Terwee CB, Bjorner JB. Estimating anchor-based minimal
important change using longitudinal confirmatory factor analysis. Qual Life Res. 2024;33:963–73.
}
