% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{squeezed_mers}
\alias{squeezed_mers}
\title{Generates all permutations of squeezed kmers}
\usage{
squeezed_mers(k = 3L)
}
\arguments{
\item{k}{kmer length}
}
\value{
vector of squeezed kmers
}
\description{
Generates all permutations of squeezed kmers
}
\examples{
squeezed_mers(3)
}
