% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_reliability.R
\name{tr_reliability}
\alias{tr_reliability}
\title{Reliability for the Transition Rates}
\usage{
tr_reliability(data, model, modification = TRUE)
}
\arguments{
\item{data}{data.frame with variables that are in \code{model}}

\item{model}{the model described using the \code{lavaan} model syntax. see model
for the examples or \code{?lavaan::model.syntax}}

\item{modification}{logical indicator if the modification index for the
latent variables should be returned. if \code{modification = FALSE} only the
reliability of the transition rating is returned.}
}
\value{
\code{list} with reliability and modification index. if
\code{modification = FALSE} only reliabilty is returned.
}
\description{
Reliability for the Transition Rates
}
\examples{
model <- '
# Factors
F1 =~ v1_1+v1_2+v1_3+v1_4+v1_5+v1_6+v1_7+v1_8+
  v1_9+v1_10+trat
F2 =~ v2_1+v2_2+v2_3+v2_4+v2_5+v2_6+v2_7+v2_8+
  v2_9+v2_10+trat

# Correlated errors over time
v1_1 ~~ v2_1
v1_2 ~~ v2_2
v1_3 ~~ v2_3
v1_4 ~~ v2_4
v1_5 ~~ v2_5
v1_6 ~~ v2_6
v1_7 ~~ v2_7
v1_8 ~~ v2_8
v1_9 ~~ v2_9
v1_10 ~~ v2_10
'
tr_reliability(data = example, model = model)
}
